refactor: break up extension file
