+++ src/extension.ts
@@ -6,74 +6,10 @@ import {
   ServerOptions,
 } from "vscode-languageclient/node";
 import { builtinFunctionDefinitions } from "./functionDefinitions";
-
-const G_FUNCTION_DEFINITION =
-  /(\bfunction\b)[\t ]*([a-zA-Z][\w]*)[\t ]*\(([^\(\)]*)\)/g; // keep in sync with syntax file rascript.tmLanguage.json #function-definitions regex
-const G_CLASS_DEFINITION = /(\bclass\b)[\t ]*(\w+)/g; // keep in sync with syntax file rascript.tmLanguage.json #class-definitions regex
-const G_COMMENTS = new RegExp("^//.*$", "g");
-
-const G_BLOCK_COMMENTS_START = /^.*\/\*.*$/g;
-const G_BLOCK_COMMENTS_END = /^.*\*\/$/g;
-const G_VARIABLES = /(\w+)\s*=/g;
-const G_STAR_BLOCK_COMMENT = /^\*.*/g; // starts with a star
-const G_FUNTION = "function";
-const G_CLASS = "class";
+import * as parser from "./parser";
+import * as models from "./models";
 
 let client: LanguageClient;
-
-interface ClassScope {
-  start: number;
-  end: number;
-  functions: Map<string, FunctionDefinition>;
-  constructorArgs: string[];
-}
-
-interface CommentBounds {
-  start: number;
-  end: number;
-  type: string;
-  raw: string;
-}
-
-interface FunctionDefinition {
-  name: string;
-  pos: vscode.Position;
-  args: string[];
-}
-
-function newClassScope(
-  start: number,
-  end: number,
-  ...constructorArgs: string[]
-): ClassScope {
-  return {
-    start: start,
-    end: end,
-    functions: new Map<string, FunctionDefinition>(),
-    constructorArgs: constructorArgs,
-  };
-}
-
-interface ClassFunction {
-  className: string;
-  name: string;
-  pos: vscode.Position;
-  args: string[];
-}
-
-function createClassFunction(
-  className: string,
-  name: string,
-  pos: vscode.Position,
-  ...args: string[]
-): ClassFunction {
-  return {
-    className: className,
-    name: name,
-    pos: pos,
-    args: args,
-  };
-}
 
 export function activate(context: vscode.ExtensionContext) {
   const rascriptLanguageServer =
@@ -141,181 +77,6 @@ function languageServer(
     });
 }
 
-function classFilter(global: boolean, usingThis: boolean, className: string) {
-  return function (el: ClassFunction) {
-    if (global) {
-      return el.className === "";
-    } else if (usingThis) {
-      return el.className === className;
-    }
-    return el.className !== "";
-  };
-}
-
-function getWordType(
-  document: vscode.TextDocument,
-  startingOffset: number,
-  endingOffset: number
-): [boolean, boolean] {
-  const text = document.getText();
-  let fn = false;
-  let cls = false;
-
-  // check for function
-  if (text[endingOffset] === "(") {
-    fn = true;
-  }
-
-  // check for previous word being class
-  let offset = startingOffset - 1;
-  while (offset >= 0) {
-    if (text[offset] !== " " && text[offset] !== "\t" && text[offset] !== "s") {
-      break;
-    }
-    if (text[offset] === "s") {
-      const position = document.positionAt(offset);
-      const range = document.getWordRangeAtPosition(position);
-      const word = document.getText(range);
-      if (word === "class") {
-        cls = true;
-      }
-      break;
-    }
-    offset--;
-  }
-  return [fn, cls];
-}
-
-function getScope(
-  document: vscode.TextDocument,
-  startingOffset: number
-): [boolean, boolean] {
-  // Determine if this function is part of a class or global function
-  const text = document.getText();
-  let global = true;
-  let usingThis = false;
-  let offset = startingOffset - 1;
-
-  while (global && offset >= 0) {
-    if (text[offset] !== " " && text[offset] !== "\t" && text[offset] !== ".") {
-      break;
-    }
-    if (text[offset] === ".") {
-      const position = document.positionAt(offset);
-      const range = document.getWordRangeAtPosition(position);
-      const word = document.getText(range);
-      if (word === "this") {
-        usingThis = true;
-      }
-      // in here means the previous non whitespace character next to the word hovered over is a dot which is the class attribute accessor operator
-      global = false;
-      break;
-    }
-    offset--;
-  }
-  return [global, usingThis];
-}
-
-function countArgsAt(document: vscode.TextDocument, offset: number) {
-  let text = document.getText();
-  let count = 0;
-  if (text[offset] === "(") {
-    offset++;
-    while (offset < text.length) {
-      if (text[offset] === ")") {
-        break;
-      }
-      if (count === 0) {
-        count = 1;
-      } else {
-        if (text[offset] === ",") {
-          count++;
-        }
-      }
-      offset++;
-    }
-  }
-  return count;
-}
-
-function getCommentBoundsList(document: vscode.TextDocument) {
-  let text = document.getText();
-  let commentBounds: CommentBounds[] = [];
-  let inComment = false;
-  let tempStart = 0;
-  for (let i = 0; i < text.length - 1; i++) {
-    if (inComment) {
-      if (text[i] === "\n" || text === "\r") {
-        inComment = false;
-        commentBounds.push({
-          start: tempStart,
-          end: i,
-          type: "Line",
-          raw: text.slice(tempStart, i + 1),
-        });
-      }
-    } else {
-      if (text[i - 1] + text[i] === "//") {
-        inComment = true;
-        tempStart = i - 1;
-      }
-    }
-    if (i === text.length - 1 && inComment) {
-      inComment = false;
-      commentBounds.push({
-        start: tempStart,
-        end: i,
-        type: "Line",
-        raw: text.slice(tempStart, i + 1),
-      });
-    }
-  }
-  // parse different comment types seperately incase they are mixed together,
-  // the bounds between these two could overlap technically
-
-  // get bounds of block comments
-  inComment = false;
-  tempStart = 0;
-  for (let i = 1; i < text.length; i++) {
-    if (inComment) {
-      if (text[i - 1] + text[i] === "*/") {
-        inComment = false;
-        commentBounds.push({
-          start: tempStart,
-          end: i,
-          type: "Block",
-          raw: text.slice(tempStart, i + 1),
-        });
-      }
-    } else {
-      if (text[i - 1] + text[i] === "/*") {
-        inComment = true;
-        tempStart = i - 1;
-      }
-    }
-    if (i === text.length - 1 && inComment) {
-      inComment = false;
-      commentBounds.push({
-        start: tempStart,
-        end: i,
-        type: "Block",
-        raw: text.slice(tempStart, i + 1),
-      });
-    }
-  }
-  return commentBounds;
-}
-
-function inCommentBound(index: number, commentBounds: CommentBounds[]) {
-  for (let i = 0; i < commentBounds.length; i++) {
-    let bound = commentBounds[i];
-    if (index >= bound.start && index <= bound.end) {
-      return true;
-    }
-  }
-  return false;
-}
-
 function localExtension(context: vscode.ExtensionContext) {
   const definitions = vscode.languages.registerDefinitionProvider("rascript", {
     provideDefinition(document, position, token) {
@@ -323,13 +84,13 @@ function localExtension(context: vscode.ExtensionContext) {
       const range = document.getWordRangeAtPosition(position);
       const word = document.getText(range);
       const origWordOffset = document.offsetAt(position);
-      let commentBounds = getCommentBoundsList(document);
-      let classes = getClassData(text, commentBounds);
+      let commentBounds = parser.getCommentBoundsList(document);
+      let classes = parser.getClassData(text, commentBounds);
       let m: RegExpExecArray | null;
-      let functionDefinitions = new Map<string, ClassFunction[]>();
-      while ((m = G_FUNCTION_DEFINITION.exec(text))) {
+      let functionDefinitions = new Map<string, models.ClassFunction[]>();
+      while ((m = parser.G_FUNCTION_DEFINITION.exec(text))) {
         // dont parse if its in a comment
-        if (inCommentBound(m.index, commentBounds)) {
+        if (parser.inCommentBound(m.index, commentBounds)) {
           continue;
         }
         let pos = document.positionAt(m.index);
@@ -338,8 +99,8 @@ function localExtension(context: vscode.ExtensionContext) {
         var args = a.filter(function (el) {
           return el !== null && el !== "" && el !== undefined;
         });
-        let item = createClassFunction(
-          detectClass(m.index, classes),
+        let item = parser.createClassFunction(
+          parser.detectClass(m.index, classes),
           m[2],
           pos,
           ...args
@@ -362,10 +123,14 @@ function localExtension(context: vscode.ExtensionContext) {
           origOffset = document.offsetAt(range.start);
         }
         let offset = origOffset - 1;
-        const [global, usingThis] = getScope(document, origOffset);
+        const [global, usingThis] = parser.getScope(document, origOffset);
         let list = functionDefinitions.get(word) || [];
         let filteredList = list.filter(
-          classFilter(global, usingThis, detectClass(origWordOffset, classes))
+          parser.classFilter(
+            global,
+            usingThis,
+            parser.detectClass(origWordOffset, classes)
+          )
         );
         // can only link to one location, so anything that has multiple definitions wont work for code jumping
         if (filteredList.length === 1) {
@@ -384,14 +149,14 @@ function localExtension(context: vscode.ExtensionContext) {
 
   const hover = vscode.languages.registerHoverProvider("rascript", {
     provideHover(document: vscode.TextDocument, position: vscode.Position) {
-      let words = new Map<string, HoverData[]>();
+      let words = new Map<string, models.HoverData[]>();
       for (let i = 0; i < builtinFunctionDefinitions.length; i++) {
         let fn = builtinFunctionDefinitions[i];
         let comment = fn.commentDoc.join("\n");
-        let hover = newHoverText(
+        let hover = parser.newHoverText(
           fn.key,
           -1,
-          G_FUNTION,
+          parser.G_FUNTION,
           "",
           comment,
           fn.url,
@@ -407,15 +172,15 @@ function localExtension(context: vscode.ExtensionContext) {
       let text = document.getText();
       let m: RegExpExecArray | null;
       // get bounds of single line comments
-      let commentBounds = getCommentBoundsList(document);
-      let classes = getClassData(text, commentBounds);
+      let commentBounds = parser.getCommentBoundsList(document);
+      let classes = parser.getClassData(text, commentBounds);
       for (const [className, classScope] of classes) {
         let pos = document.positionAt(classScope.start);
-        let comment = getCommentText(document, pos);
-        let hover = newHoverText(
+        let comment = parser.getCommentText(document, pos);
+        let hover = parser.newHoverText(
           className,
           classScope.start,
-          G_CLASS,
+          parser.G_CLASS,
           "",
           comment,
           "",
@@ -428,22 +193,22 @@ function localExtension(context: vscode.ExtensionContext) {
           words.set(className, [hover]);
         }
       }
-      while ((m = G_FUNCTION_DEFINITION.exec(text))) {
+      while ((m = parser.G_FUNCTION_DEFINITION.exec(text))) {
         // dont parse if its in a comment
-        if (inCommentBound(m.index, commentBounds)) {
+        if (parser.inCommentBound(m.index, commentBounds)) {
           continue;
         }
-        let className = detectClass(m.index, classes);
+        let className = parser.detectClass(m.index, classes);
         let pos = document.positionAt(m.index);
-        let comment = getCommentText(document, pos);
+        let comment = parser.getCommentText(document, pos);
         let a = m[3].split(",").map((s) => s.trim());
         var args = a.filter(function (el) {
           return el !== null && el !== "" && el !== undefined;
         });
-        let hover = newHoverText(
+        let hover = parser.newHoverText(
           m[2],
           m.index,
-          G_FUNTION,
+          parser.G_FUNTION,
           className,
           comment,
           "",
@@ -468,7 +233,7 @@ function localExtension(context: vscode.ExtensionContext) {
       const startingOffset = document.offsetAt(startingPos);
       const endingOffset = document.offsetAt(endingPos);
       const word = document.getText(range);
-      const hoverClass = detectClass(startingOffset, classes);
+      const hoverClass = parser.detectClass(startingOffset, classes);
       let offset = startingOffset - 1; // get character just before the function name position
 
       // Special case: this keyword should show the class hover info
@@ -483,11 +248,15 @@ function localExtension(context: vscode.ExtensionContext) {
           }
         }
       }
-      const [global, usingThis] = getScope(document, startingOffset);
+      const [global, usingThis] = parser.getScope(document, startingOffset);
 
       let definitions = words.get(word);
       if (definitions !== undefined) {
-        const [fn, cls] = getWordType(document, startingOffset, endingOffset);
+        const [fn, cls] = parser.getWordType(
+          document,
+          startingOffset,
+          endingOffset
+        );
         if (!fn && !cls) {
           // only provide hover data for classes and functions
           return null;
@@ -522,7 +291,7 @@ function localExtension(context: vscode.ExtensionContext) {
               // if its a function, further filter down by arg list length
               // otherwise just append if its a class
               if (fn) {
-                let numArgs = countArgsAt(document, endingOffset);
+                let numArgs = parser.countArgsAt(document, endingOffset);
                 if (numArgs === definition.args.length) {
                   filteredDefinitions.push(definition);
                 }
@@ -566,23 +335,23 @@ function localExtension(context: vscode.ExtensionContext) {
           completionFunctions.push(fn.key);
         }
         let text = document.getText();
-        let commentBounds = getCommentBoundsList(document);
-        let classes = getClassData(text, commentBounds);
+        let commentBounds = parser.getCommentBoundsList(document);
+        let classes = parser.getClassData(text, commentBounds);
         let m: RegExpExecArray | null;
-        while ((m = G_FUNCTION_DEFINITION.exec(text))) {
+        while ((m = parser.G_FUNCTION_DEFINITION.exec(text))) {
           // dont parse if its in a comment
-          if (inCommentBound(m.index, commentBounds)) {
+          if (parser.inCommentBound(m.index, commentBounds)) {
             continue;
           }
           completionFunctions.push(m[2]);
         }
         let functionSet: Set<string> = new Set(completionFunctions);
         functionSet.forEach((fnName: string) => {
-          completionItems.push(newBuiltInFunction(fnName));
+          completionItems.push(parser.newBuiltInFunction(fnName));
         });
-        while ((m = G_VARIABLES.exec(text))) {
+        while ((m = parser.G_VARIABLES.exec(text))) {
           // dont parse if its in a comment
-          if (inCommentBound(m.index, commentBounds)) {
+          if (parser.inCommentBound(m.index, commentBounds)) {
             continue;
           }
           completionVariables.push(m[1]);
@@ -590,7 +359,7 @@ function localExtension(context: vscode.ExtensionContext) {
         let variableSet: Set<string> = new Set(completionVariables);
         variableSet.forEach((varName: string) => {
           completionItems.push(
-            newCompletion(varName, vscode.CompletionItemKind.Variable)
+            parser.newCompletion(varName, vscode.CompletionItemKind.Variable)
           );
         });
         for (const [className, classScope] of classes) {
@@ -599,7 +368,7 @@ function localExtension(context: vscode.ExtensionContext) {
         let classSet: Set<string> = new Set(completionClasses);
         classSet.forEach((className: string) => {
           completionItems.push(
-            newCompletion(className, vscode.CompletionItemKind.Class)
+            parser.newCompletion(className, vscode.CompletionItemKind.Class)
           );
         });
         return completionItems;
@@ -615,285 +384,4 @@ export function deactivate(): Thenable<void> | undefined {
     return undefined;
   }
   return client.stop();
-}
-
-function getCommentText(
-  document: vscode.TextDocument,
-  pos: vscode.Position
-): string {
-  let comment = "";
-  let untrimmedComment = ""; // This holds a second copy of the comments with leading stars
-  let blockCommentStarStyle = true;
-  if (pos.line > 0) {
-    // dont look for comments if were at the top of the file
-    let offset = 1;
-    let inBlock = false;
-    // while not at the top of the file and the next line up is a comment
-    while (pos.line - offset >= 0) {
-      let lineNum = pos.line - offset;
-      let line = document.lineAt(new vscode.Position(lineNum, 0)).text;
-      line = line.trimStart();
-      if (offset === 1) {
-        // if were right above the function declaration, look for a block comment
-        let isBlock = G_BLOCK_COMMENTS_END.test(line);
-        G_BLOCK_COMMENTS_END.lastIndex = 0;
-        if (isBlock) {
-          inBlock = true;
-        }
-      }
-      if (inBlock) {
-        // handle block comments
-        let endBlock = G_BLOCK_COMMENTS_START.test(line);
-        G_BLOCK_COMMENTS_START.lastIndex = 0;
-        if (endBlock) {
-          // at the beginning of comment block
-
-          // TRIM START TOKEN
-          let trimmedLine = line.split(/\/\*(.*)/s); // use whats after the start token
-          let firstEl = trimmedLine.shift(); // remove text before commend block start
-          let newLine = trimmedLine.join("").trimStart();
-
-          // TRIM END TOKEN
-          trimmedLine = newLine.split("*/"); // use whats after the star token
-          if (trimmedLine.length > 2) {
-            let firstEl = trimmedLine.pop();
-          }
-          newLine = trimmedLine.join("").trimStart();
-          if (blockCommentStarStyle) {
-            let starComment = G_STAR_BLOCK_COMMENT.test(newLine);
-            G_STAR_BLOCK_COMMENT.lastIndex = 0;
-            if (!starComment) {
-              blockCommentStarStyle = false;
-            }
-          }
-          untrimmedComment = "//" + newLine + "\n" + untrimmedComment; // keep an untrimmed version of the comment in case the entire block is prefixed with stars
-
-          // TRIM FIRST '*' TOKEN (in case they comment that way)
-          trimmedLine = newLine.split(/^\*(.*)/s); // use whats after the end token
-          if (trimmedLine.length > 2) {
-            let firstEl = trimmedLine.shift(); // remove leading '*'
-          }
-          newLine = trimmedLine.join("").trimStart();
-          comment = "//" + newLine + "\n" + comment;
-          break;
-        } else {
-          // at end of comment block
-
-          // TRIM END TOKEN (guaranteed to not have text after end token if tthe user wants comments to appear in hover box)
-          let trimmedLine = line.split("*/"); // use whats after the end token
-          if (trimmedLine.length > 2) {
-            let firstEl = trimmedLine.pop();
-          }
-          let newLine = trimmedLine.join("").trimStart();
-
-          if (blockCommentStarStyle) {
-            let starComment = G_STAR_BLOCK_COMMENT.test(newLine);
-            G_STAR_BLOCK_COMMENT.lastIndex = 0;
-            if (!starComment) {
-              blockCommentStarStyle = false;
-            }
-          }
-          untrimmedComment = "//" + newLine + "\n" + untrimmedComment; // keep an untrimmed version of the comment in case the entire block is prefixed with stars
-
-          // TRIM FIRST '*' TOKEN (in case they comment that way)
-          trimmedLine = newLine.split(/^\*(.*)/s); // use whats after the first star token
-          if (trimmedLine.length > 2) {
-            let firstEl = trimmedLine.shift(); // remove leading '*'
-          }
-          newLine = trimmedLine.join("").trimStart();
-          comment = "//" + trimmedLine[0] + "\n" + comment;
-        }
-      } else {
-        // else handle single line block comments
-        let isComment = G_COMMENTS.test(line);
-        G_COMMENTS.lastIndex = 0;
-        if (isComment) {
-          comment = line + "\n" + comment;
-        } else {
-          break;
-        }
-      }
-      offset = offset + 1;
-    }
-  }
-  let finalComment = untrimmedComment;
-  if (blockCommentStarStyle) {
-    finalComment = comment;
-  }
-  return finalComment;
-}
-
-function detectClass(funcPos: number, classData: Map<string, ClassScope>) {
-  for (const [className, classScope] of classData) {
-    if (funcPos >= classScope.start && funcPos <= classScope.end) {
-      return className;
-    }
-  }
-  return "";
-}
-
-function getClassData(text: string, commentBounds: CommentBounds[]) {
-  let classes = new Map<string, ClassScope>();
-  let m: RegExpExecArray | null;
-  while ((m = G_CLASS_DEFINITION.exec(text))) {
-    // dont parse if its in a comment
-    if (inCommentBound(m.index, commentBounds)) {
-      continue;
-    }
-    let postClassNameInd = m.index + m[0].length;
-    let ind = postClassNameInd;
-    let stack = []; // makeshift stack to detect scope of class
-    let strippedText = ""; // this is used to determine the implicit arguments to a class constructor
-    while (ind < text.length) {
-      // anything other than white space or open curly brace is an error and we just wont parse this class
-      if (
-        text[ind] !== " " &&
-        text[ind] !== "\n" &&
-        text[ind] !== "\r" &&
-        text[ind] !== "\t" &&
-        text[ind] !== "{"
-      ) {
-        break;
-      }
-      if (text[ind] === "{") {
-        // get the position of the opening curly brace
-        stack.push(ind);
-        break;
-      }
-      ind++;
-    }
-    if (stack.length === 1) {
-      // if we have a curly brace scope, start parsing to find the end of the scope
-      let ind = stack[0] + 1; // next char after our first open curly brace
-      while (ind < text.length) {
-        if (text[ind] === "}") {
-          stack.pop();
-        } else if (text[ind] === "{") {
-          stack.push(ind);
-        } else {
-          let size = stack.length;
-          if (size === 1) {
-            // if the code is at the first level of the class (not in a function) append it to our stripped class
-            strippedText = strippedText + text[ind];
-          }
-        }
-        let size = stack.length;
-        if (size === 0) {
-          // we have found our end position of the scope, break out
-          break;
-        }
-        ind++;
-      }
-      let args = [];
-      let m2: RegExpExecArray | null;
-      while ((m2 = G_VARIABLES.exec(strippedText))) {
-        args.push(m2[1]);
-      }
-      let scope = newClassScope(m.index, ind, ...args);
-      classes.set(m[2], scope);
-    }
-  }
-  return classes;
-}
-
-function newBuiltInFunction(name: string) {
-  const snippetCompletion = new vscode.CompletionItem(name);
-  snippetCompletion.insertText = new vscode.SnippetString(name + "()");
-  snippetCompletion.kind = vscode.CompletionItemKind.Function;
-  const moveCursorCommand: vscode.Command = {
-    title: "Move cursor left between parentheses",
-    command: "cursorLeft",
-  };
-
-  snippetCompletion.command = moveCursorCommand;
-  return snippetCompletion;
-}
-
-function newCompletion(name: string, kind: vscode.CompletionItemKind) {
-  const snippetCompletion = new vscode.CompletionItem(name, kind);
-
-  return snippetCompletion;
-}
-
-interface HoverData {
-  key: string;
-  index: number;
-  className: string;
-  hover: vscode.Hover;
-  args: string[];
-}
-
-function newHoverText(
-  key: string,
-  index: number,
-  type: string,
-  className: string,
-  text: string,
-  docUrl: string,
-  ...args: string[]
-): HoverData {
-  let argStr = args.join(", ");
-  let commentLines = text.split(/\r?\n/);
-  let lines = [];
-  let prefix = "function ";
-  if (className !== "") {
-    prefix = `// class ${className}\nfunction `;
-  }
-  lines.push(`\`\`\`rascript\n${prefix}${key}(${argStr})\n\`\`\``);
-  if (type === G_CLASS) {
-    let fnLine = lines[0];
-    lines = [`\`\`\`rascript\nclass ${key}\n\`\`\``];
-    lines.push(fnLine);
-  }
-  if (text !== "") {
-    lines.push("---");
-    let curr = "";
-    let codeBlock = false;
-    for (let i = 0; i < commentLines.length; i++) {
-      let line = commentLines[i].replace(/^\/\//g, "");
-      line = line.trimStart();
-      if (line.startsWith("```")) {
-        codeBlock = !codeBlock;
-        if (codeBlock) {
-          curr = line;
-        } else {
-          curr = curr + "\n" + line;
-          lines.push(curr);
-          curr = "";
-        }
-        continue;
-      }
-      if (line.startsWith("|") || line.startsWith("*")) {
-        line = line + "\n";
-      }
-      if (codeBlock) {
-        curr = curr + "\n" + line;
-      } else {
-        if (line === "") {
-          lines.push(curr);
-          curr = "";
-        } else {
-          curr = curr + " " + line;
-        }
-      }
-    }
-    if (curr !== "") {
-      lines.push(curr);
-    }
-    if (codeBlock) {
-      lines.push("```");
-    }
-  }
-  if (docUrl !== "") {
-    lines.push("---");
-    lines.push(`[Wiki link for \`${key}()\`](${docUrl})`);
-  }
-
-  return {
-    key: key,
-    index: index,
-    className: className,
-    hover: new vscode.Hover(lines),
-    args: args,
-  };
 }
